+++ tests/testthat/test-hit_api.R
@@ -7,9 +7,9 @@ test_that("build_req works", {
    expect_equal(res$url, "https://apicarto.ign.fr/path_to_test?param1=param1")
 
    # error case
-   expect_error(build_req(path = 1234), "but has class 'numeric'")
-   expect_error(build_req(path = NA), "but has class 'logical'")
-   expect_error(build_req(path = NULL), "but has class 'NULL'")
+   expect_error(build_req(path = 1234), "must be of class character")
+   expect_error(build_req(path = NA), "must be of class character")
+   expect_error(build_req(path = NULL), "must be of class character")
    expect_error(build_req(path, "test"), "All components of ... must be named")
 })